refactor: merge StartPlaying screen into WorldPreGeneration screen
