--- engine/src/main/resources/org/terasology/engine/assets/ui/menu/startPlayingScreen.ui
@@ -1,153 +1,0 @@
-{
-  "type": "engine:StartPlayingScreen",
-  "skin": "engine:mainMenu",
-  "contents": {
-    "type": "relativeLayout",
-    "contents": [
-      {
-        "type": "UILabel",
-        "id": "title",
-        "family": "title",
-        "text": "${engine:menu#start-playing}",
-        "layoutInfo": {
-          "height": 48,
-          "position-horizontal-center": {},
-          "position-top": {
-            "target": "TOP",
-            "offset": 48
-          }
-        }
-      },
-      {
-        "type": "UILabel",
-        "id": "subtitle",
-        "layoutInfo": {
-          "use-content-height": true,
-          "use-content-width": true,
-          "position-horizontal-center": {},
-          "position-top": {
-            "target": "BOTTOM",
-            "widget": "title",
-            "offset": 32
-          }
-        }
-      },
-      {
-        "type": "UIBox",
-        "id": "container",
-        "layoutInfo": {
-          "position-horizontal-center": {},
-          "width": 392,
-          "position-top": {
-            "target": "TOP",
-            "offset": -8,
-            "widget": "coreLayout"
-          },
-          "position-bottom": {
-            "target": "BOTTOM",
-            "offset": -8,
-            "widget": "coreLayout"
-          }
-        }
-      },
-      {
-        "type": "ColumnLayout",
-        "id": "coreLayout",
-        "columns": 1,
-        "horizontalSpacing": 8,
-        "layoutInfo": {
-          "width": 384,
-          "use-content-height": true,
-          "position-horizontal-center": {},
-          "position-vertical-center": {}
-        },
-        "contents": [
-          {
-            "type": "UIImage",
-            "skin": "framed_image",
-            "id": "preview"
-          }
-        ]
-      },
-      {
-        "type": "UIButton",
-        "text": "${engine:menu#return-pregeneration}",
-        "id": "close",
-        "layoutInfo": {
-          "width": 120,
-          "height": 80,
-          "position-vertical-center": {
-            "target": "CENTER",
-            "widget": "container"
-          },
-          "position-right": {
-            "target": "LEFT",
-            "widget": "container",
-            "offset": 16
-          }
-        }
-      },
-      {
-        "type": "RowLayout",
-        "id": "actionsRow",
-        "horizontalSpacing": 32,
-        "contents": [
-          {
-            "type": "UIButton",
-            "text": "${engine:menu#return-main-menu}",
-            "id": "mainMenu"
-          },
-          {
-            "type": "UIButton",
-            "text": "${engine:menu#start-playing}",
-            "id": "play"
-          }
-        ],
-        "layoutInfo": {
-          "width": 350,
-          "height": 80,
-          "position-horizontal-center": {},
-          "position-top": {
-            "target": "BOTTOM",
-            "widget": "container",
-            "offset": 16
-          }
-        }
-      },
-      {
-        "type": "UIImage",
-        "image": "engine:happyGooey",
-        "layoutInfo": {
-          "width": 87,
-          "height": 40,
-          "position-left": {
-            "target": "RIGHT",
-            "widget": "actionsRow",
-            "offset": 8
-          },
-          "position-vertical-center": {
-            "target": "CENTER",
-            "widget": "actionsRow"
-          }
-        }
-      },
-      {
-        "type": "UIImage",
-        "image": "engine:angryGooey",
-        "layoutInfo": {
-          "width": 87,
-          "height": 40,
-          "position-right": {
-            "target": "LEFT",
-            "widget": "actionsRow",
-            "offset": 8
-          },
-          "position-vertical-center": {
-            "target": "CENTER",
-            "widget": "actionsRow"
-          }
-        }
-      }
-    ]
-  }
-}
